test: change and add several test
After refactoring update and implement test.
